[plan-2026-08-30-3e8a6ff3/iter-1/step-2] dtype: deep_kernel_pca follows the active policy

+++ src/dl_techniques/layers/statistics/deep_kernel_pca.py
@@ -711,7 +711,15 @@ class DeepKernelPCA(keras.layers.Layer):
 
         # Add regularization to diagonal for numerical stability.
         # ops.eye accepts the symbolic batch size under TF graph mode (verified).
-        kernel_matrix_reg = kernel_matrix + self.regularization_lambda * ops.eye(batch_size)
+        # DECISION plan-2026-08-30T175846-3e8a6ff3/D-002
+        # Take the identity's dtype from `kernel_matrix`. Do NOT write bare
+        # `ops.eye(batch_size)` nor a literal `dtype="float32"`: both pin
+        # backend.floatx(), and this line then raised `cannot compute AddV2 ...
+        # expected to be a half/bfloat16 tensor but is a float tensor` under
+        # mixed_float16 and mixed_bfloat16 (measured). decisions.md D-002.
+        kernel_matrix_reg = kernel_matrix + self.regularization_lambda * ops.eye(
+            batch_size, dtype=kernel_matrix.dtype
+        )
 
         # Keep the projection columns on the unit sphere during training. Test
         # `training is True`, not truthiness: a symbolic or None flag under
@@ -1165,8 +1173,14 @@ class DeepKernelPCA(keras.layers.Layer):
             # Weight features by their level (deeper levels get slightly less weight)
             weighted_features = []
             for i, features in enumerate(output_features):
-                # Exponential decay weighting
-                weight = ops.exp(-0.1 * i)
+                # Exponential decay weighting.
+                # DECISION plan-2026-08-30T175846-3e8a6ff3/D-002
+                # `i` is a Python int, so `ops.exp(-0.1 * i)` alone materializes
+                # the weight at backend.floatx() and the multiply below raised
+                # `cannot compute Mul ... expected to be a float tensor but is a
+                # half/double/bfloat16 tensor` (measured). Source the dtype from
+                # `features`, not from a literal. decisions.md D-002.
+                weight = ops.exp(ops.cast(-0.1 * i, features.dtype))
                 weighted_features.append(weight * features)
 
             # Concatenate weighted features
